Bug fixes and linting changes

+++ shiftleft-utils/bulk_delete_projects.py
@@ -39,12 +39,11 @@ def _delete_project(cookie, resource_url):
     First we issue a `DELETE` on the resource and retrieve a confirmation token, then we issue a second `DELETE` on the
     same resource adding the token as the body.
     """
-    print(F"Will initiate deletion of {resource_url}")
+    print(f"Will initiate deletion of {resource_url}")
     headers = CaseInsensitiveDict()
     headers["Cookie"] = cookie
-    res = requests.delete(url=resource_url,
-                          headers=headers)
-    print(F"got response {res.text}")
+    res = requests.delete(url=resource_url, headers=headers)
+    print(f"got response {res.text}")
     if res.status_code != 200:
         raise ErrReqFailed(res.status_code)
 
@@ -52,11 +51,9 @@ def _delete_project(cookie, resource_url):
 
     confirmation_token = json.dumps({"response": res_map})
 
-    print(F"Will confirm deletion of {resource_url} with token {confirmation_token}")
+    print(f"Will confirm deletion of {resource_url} with token {confirmation_token}")
     headers["content-type"] = "application/json"
-    res = requests.delete(url=resource_url,
-                          headers=headers,
-                          data=confirmation_token)
+    res = requests.delete(url=resource_url, headers=headers, data=confirmation_token)
     if res.status_code != 200:
         raise ErrConfirmationReqFailed(res.status_code)
 
@@ -68,7 +65,10 @@ def delete_project(cookie, org_id, project_id=None):
     if project_id is None:
         print("received an empty project ID, this deletion will not take place")
         return
-    _delete_project(cookie, F"https://app.shiftleft.io/api/v2/organizations/{org_id}/projects/{project_id}".strip())
+    _delete_project(
+        cookie,
+        f"https://app.shiftleft.io/api/v2/organizations/{org_id}/projects/{project_id}".strip(),
+    )
 
 
 def bulk_delete_projects():
@@ -78,14 +78,14 @@ def bulk_delete_projects():
     """
     cookie = os.environ.get(SHIFTLEFT_COOKIE_ENV_VAR_NAME, None)
     if cookie is None:
-        sys.exit(F"{SHIFTLEFT_COOKIE_ENV_VAR_NAME} {PLS_SET_ERR}")
+        sys.exit(f"{SHIFTLEFT_COOKIE_ENV_VAR_NAME} {PLS_SET_ERR}")
     org_id = os.environ.get(SHIFTLEFT_ORG_ID_ENV_VAR_NANE, None)
     if org_id is None:
-        sys.exit(F"{SHIFTLEFT_ORG_ID_ENV_VAR_NANE} {PLS_SET_ERR}")
+        sys.exit(f"{SHIFTLEFT_ORG_ID_ENV_VAR_NANE} {PLS_SET_ERR}")
 
     project_list_file_name = abspath(sys.argv[1])
     if not exists(project_list_file_name):
-        sys.exit(F"cannot find the specified project id list: {project_list_file_name}")
+        sys.exit(f"cannot find the specified project id list: {project_list_file_name}")
 
     with open(project_list_file_name) as project_file:
         lines = project_file.readlines()
@@ -95,13 +95,17 @@ def bulk_delete_projects():
             try:
                 delete_project(cookie, org_id, line)
             except ErrReqFailed as e:
-                print(F"Failed to initiate deletion for project: {line} with HTTP code: {e.code}")
+                print(
+                    f"Failed to initiate deletion for project: {line} with HTTP code: {e.code}"
+                )
                 with open(FAILED_DELETES_FILE, "a") as f:
-                    print(F"{line}\n", file=f)
+                    print(f"{line}\n", file=f)
             except ErrConfirmationReqFailed as e:
-                print(F"Failed to complete deletion for project: {line} with HTTP code: {e.code}")
+                print(
+                    f"Failed to complete deletion for project: {line} with HTTP code: {e.code}"
+                )
                 with open(FAILED_DELETES_FILE, "a") as f:
-                    print(F"{line}\n", file=f)
+                    print(f"{line}\n", file=f)
 
 
 if __name__ == "__main__":
